Complex Topology Tests
Test 105 now behaving as expected (this is the test for use of an ExtraJob invariant that uses the wrong value.
Also noted that the function AuditEvent::ProcessAudit Events does not currentlty test for invalid values (this is left to the state AssessingInvariants) but if added would mean that we could fail the Job much earlier when this is detected. In conversations with Colin we decided that the additional complexity need (though minor) outweighed the benefits at this stage - so we haveadded a TODO to review this decision.