job:  #MUN-152  support loops in play
Loops are supported when playing a sequence.  The logic is like this:
When exactly 2 events follow the current event, the index of each is
checked to see if it is lower than the current event (thus earlier
in the sequence).  This lower event is selected as the next event,
and the other next event is omitted (for now).  After an arbitrary
loop count (of 4), the "carry on" event is selected.