Complex topology tests
New scenario added: InitCompleEventSequence2Definition that sets up 4 new Job definitions for testing various BRANCH and MERGE scenarios.
Test73 added to show that a number of sequence variants are valid for a single job definition if it does not include MERGECOUNT controls to constrain the topology.
Some improvements to log messages