Remove inaccessible MASL from AuditEvent,AddEventForNewJob
This change to isSequenceStart=true had already been made. So I removed the test for the same condition that followed, leaving the logMessage part and deleting the inaccessible else part. I also checked that the same problem was not present in JobInProgress when a start event for an additional sequence within the Job is seen. That's fine - no changes needed.