job:  #MUN-167  base for release video
Greg did the heavy lifting.  This all works.  I took out the
connection to the Inspector from verifier.yml.  This allowed
the default stdout to show more easily.  I removed the old-style
sequence diagram .puml file.  We may want to bring back the job_definitions
folder in the future.
There is more clean-up to be done and .gitignore files to be added,
but I am merging this now so others can see and use it.  And it is
technically good enough for the release.