job:  #MUN2-14  AEO _event_data load workaround
When AEO reads/updates the system spec, it runs into a forward reference.
The reason is that separate event data files are (now) linked by a shared
extra-job token.  If the user of the shared item loads ahead of the source,
the spec will not find it.
The workaround is to load these files in alphabetical order, meaning that
we need to name Source files alphabetically ahead of User files (convenient
for 'Source' and 'User' in the name).