Updated AER and AEO so that jobName is passed with the event
It has now been agreed that jobName will be passed in the event data with each event. AER has now been updated to pass jobName to AEO.

AEO now uses the supplied jobName and validates the supplied jobName and the event against the jobType and the eventType. If the supplied jobName is not valid for the event it shall fail.

The jobName applied to the job is dependant on the first event the EO receives. If subsequent events for the same jobId have a different jobName then this shall cause a job failure.

Updated the projects to reflect the change in the interface between AER and AEO.