Fixed bug identified in event type creation
When creating the audit event type AEO needs to check if the event currently exists in the job definition but was checking for the event type exisiting across all job definitions.

In doing the change for providing the jobName to SVDC the search for the existing event type was not constrained to only events in that specific job.
When we have two job definitions that contain an event type with the same name when AEO was processing the second job definition it was finding an event type from the first job definition which resulted in the event type not being created.

AEO has been updated to check for event types in the job definition that is being processed.