Updates and new unit tests for SVDC re repeated event names
The reuse of event names in different jobs is now supported by the model. This commit hardens the invariant and dynamic controls parts of SVDC to tolerate the reuse of event names. This includes a unit test of a complex scenario in which a new job definition is added that reuses all of the event names of an early job. This now passes. The upgraded software has been tested end to end using regression, unhappy_alarms and unhappy_successes. All tests pass.