Updates for metrics and config definitions
The additional alarms have been added to the config.yml in the grok exporter.

A technical not has been created to capture the configuration items that are allowed in the pv-config.json.

The pv-config.json has been updated to remove redundant configuration items.

A correction was allied to the reading of the stale audit event configuration data in AEOrdering.