Updated to run ISTORE persistently
Updated to build IStore persistently. Please note that as there is only one conanfile.py for PV_PROC and I want ISTORE_PROC to be build persistently I have had to turn on the persistent build for everything, e.g. all domains and all processes. @leviathan747 can you please use your conan skills to update how the build works in PV_PROC as the only things that need to be built persistently in that model directory is ISTORE_PROC and the consequence of this is that it is slowing down the build.

Updated the execution of IStore to read in a command line argument at start up '-fileStore' and if present persist invariants to file. This should only be used when the domain is run transiently.

Added command line argument '-fileStore' to command line for PV_PROC in the single process execution defined in delpoy/docker-compose.yml where IStore is part of PV_PROC which is executed transiently. This ensures that the invariants are persisted to file.

Updated the deploy/docker-compose.kafka.yml to execute the persistent version of IStore. This currently has the additional command line argument of -cold which will delete and create a new database every time it is executed. The reason for the -cold flag is that there is an outstanding bug against the architecture to ensure that if the persistent process is run without the -cold command line option and the database does not exist then it should be created https://github.com/xtuml/masl/pull/32.