Fix for completion test which was incorrect of an AND branch was not taken
This update adds some logic to the test for sequence/job completion which was incorrect if no event an AND branch was seen before the other branch(es) completed. A test has been added to look for any AND fork that has been touched with at least one fork to ensure that all branches are present before concluding the job is complete. This means the AND completeness test in a later state is not needed any more. An incomplete AND fork will result in a job timing out because its waiting for that additional event.