Completed Job Management work
Job Management is now completed. There is a new design document in notes MUN2-147-horizontal-scaling-design.adoc

All the Job Management tests pass, the regression tests all pass and the benchmark tests run.

The time for PV to complete in the regression script has been extended as the tests were validated before they were complete.

I have commented out some of the lines in the benchmark python file as it was not capturing the new events for job management. This is because Job Management does not report the event id in its benchmark data as it does not recover the event id from the incoming string. This python file will need to be updated to make use of the new reports that Job Management is providing as described in the design document.