Updated logging
Job Management:
Removed event received and event written logs. Added a new log for heartbeat failure and logs for deregister.

Added new state Working after AvailableForWork to EmployedWorker. Added working attribute to EmployedWorker which is false when not registered and gets set to true on first assignment of work. Moved first assignment of work from when Job Management has registered the worker to when JobManagment has received the first heartbeat. The registerWorker service was deregistering an employed worker if it received a register request when the worker was already registered. This was so any assigned jobs could be re-assigned as the worker did not think it was registered. This will now only happen if the employed worker has sent the first heartbeat and the working attribute is true.

AEOrdering: Fixed type in log. Added log for deregistered request. Cancelled timer when NotRegistered state entered.