Updates to fix integration issues
Fixed issue with retirement data for worker not being ste.

Removed the periodic writing of completed Job Ids to a file. Job Ids are now stored for a configurable time and then deleted.

Put jobs being processed by a worker that becomes retired to front of queue.

Reset the heartbeat counter when the jobComplete has been received.

Added logging fro when a worker becomes registered.