job:  #188  initial move to new loggers
This commit contains a logger configuration file that defines new
loggers according to the documented strategy.  Only one change to
the models has been made; Verification Gateway is now using the
pv.vg(.verifiablejob) debug logger.  And it works.
A new MASL with support for a Kafka log appender is the architecture.
It was discovered that the log config was inconsistent with the
docker compose.  The loggers were configured to run at DEBUG level,
but only AsyncLogger was using the configuration.  Soon, we will need
to split the log configuration apart to have production and development
versions.