job:  #188  begin migration to new logging
This commit begins using the new logging/reporting strategy.
AsyncLogger is deprecated.  It has been taken out of the docker
compose file, but it has not been archived from the repository.
Messages that used to go through AsyncLogger are using Logger.

Verification Gateway is logging the audit events (for PVprime)
using the new logging channel.  This channel will eventually be
a kafka appender.  But there may be work in getting the format
to line up with that expected by Reception.

The BenchmarkingProbe is commented out for now.  Instead a Logger
with a Kafka appender is being used.  Again, formatting is a question.