job:  #188  log filtering and job management
In this commit there is the filtering of logs when running DEBUG level.
Everything INFO and above goes to the normal reporting log.  Everything
DEBUG and below goes to the debug log.  The same logger but different
levels go to different appenders.

Job Management now joins Verification Gateway in using the mechanics
of the new logging strategy.  Log messages are still not structured.