job:  #188  Add event counts to Job Management
Keep a running count of audit events received by the job manager.
Also keep a running count of the events reported to each worker.
The worker event counts should sum up to the total received count.
The counts among the workers should also be relatively (not exactly)
alanced.  On the dashboard, if these begin to skew, it will be
indication of a problem.