job:  #212  fix race condition
SVDC requests restoration of an invariant from IStore.  IStore responds
immediately and gets an event queued to SVDC ahead of the event that is
to put SVDC into the wait state (waiting on Persistence).
The solution initially proposed here it to use an accelerated event to self
in place of the polymorphic event.
An alternative is to add a transition between Assessing and Reassessing.