job:  #213  deleting AEReception and AsyncLogger
The AEReception model and all artefacts are now deprecated and deleted.
The AsyncLogger model and all artefacts are now deprecated and deleted.
References in documentation have been updated.
The docker compose file for production needed updates for AsyncLogger,
but it also had not been modified for the changes in Kafka topics.  This
has now been done in this branch.