job:  #224  fix bug in JM/AEO bring-up
In a conversation, I learned that JM is expecting to begin assigning
work to an EmployedWorker only after receiving the first heartbeat.
This was not happening.  A small change to JM to include 'working'
in the selection of a worker for an assignment together with a small
change in AEO to provide a first heartbeat upon successful registration.
We seem to be synchronised now.