job:  #226  start/stop/start config
This commit supports starting the PV in the face of a pre-populated
Kafka message topic.  It also support stopping and restarting the PV.
The PV now "picks up where it left off".  Such a restart does result
in a few job failures, but a surprising few.  The jobs that fail are
those that were mid-flight processing when the PV came down.