job:  #154  step 1 Job Management JSON
Work with native JSON audit events in Job Management.  This is a first
step toward handling arrays of audit events coming from the message
broker.  The biggest change is AuditEvent.acceptAuditEvents (note plural)
which can receive either a single audit event JSON object or a JSON array
of audit events.  Most of the new code is additive for now.  Deprecated
code will be removed in the next steps.