job:  #154  support arrays of JSON audit events
This commit adds the capability to exchange JSON arrays of audit
event between JM and AEO.  AEO now has AcceptJsonAuditEvents which
takes the workerId and a native JSON::Array of audit events for a
single job.  This array may or may not be sorted; AEO does the Right
Thing in both cases.
This commit sees performance more than doubled.  This performance
increase is experienced even if the events in the message/job are
shuffled.  It seems that the performance increase is due primarily
to the reduced chatter with the message broker.
Experimentation will next try to deal with input arrays of events
at JM with mixed jobs.