job:  #154 #224  JM unit tests
Issue #224 introduced an extra handshake during the bring-up of PV.
It added a heartbeat message from AEO immediately upon worker registration.
This avoided a delay in employing workers.  It was a good change, but it
broke the unit tests.  This commit fixes the unit tests by adding the
heartbeat response directly after registering a worker.