job:  #246  support single event sequences
This commit add support for sequences with a single event.
It is noted that single event sequences represent very loose
constraints on a job and represent potentially dubious protocol
verification (virtually no protocol).
SequenceVerification is updated to deal with events which are both
a start event and and end event.
Test cases are added.