job:  #262  throttle throttling a little
This change reduces the amount of interaction with the message credit
mechanism.  Through configuration it can be mostly turned off (by having
a very high MaxUnassignedJobs).  An Info level log message is delivered
each time throttling is enabled or disabled.  This will never be seen if
throttling never commences in the first place.